+++ hexagonal/src/main/java/com/example/hexagonal/adapters/in/controller/response/CustomerResponse.java
@@ -1,0 +1,14 @@
+package com.example.hexagonal.application.ports.in;
+
+import com.example.hexagonal.adapters.in.controller.response.AddressResponse;
+import lombok.Data;
+
+@Data
+public class CustomerResponse {
+
+    private String name;
+    private AddressResponse address;
+    private String cpf;
+    private Boolean isValidCpf;
+
+}