+++ hexagonal/src/main/java/com/example/hexagonal/adapters/out/repository/mapper/CustomerEntityMapper.java
@@ -7,5 +7,7 @@ import org.mapstruct.Mapper;
 @Mapper(componentModel = "spring")
 public interface CustomerEntityMapper {
 
-    CustomerEntity toCustomer(Customer customer);
+    CustomerEntity toCustomerEntity(Customer customer);
+
+    Customer toCustomer(CustomerEntity customerEntity);
 }