+++ CHANGELOG.md
@@ -10,3 +10,5 @@
     - Creating Controller.
 - CRUD - Read
     - Creating UseCase.
+    - Creating Adapter.
+    - Creating Controller.